requirements.txt
websocket_client==0.57.0
requests==2.23.0
Shapely==1.7.0
geojson==2.5.0
